// frontend/src/api/generated/tasks.ts
export interface RestTaskResponse {
  action_item_completed_count: number;
  action_item_count: number;
  /** @nullable */
  actual_minutes: number | null;
  assignee_id: string;
  can_update: boolean;
  created_at: string;
  description: string;
  /** @nullable */
  due_date: string | null;
  estimate_source: string;
  /** @nullable */
  estimated_minutes: number | null;
  id: string;
  /** @nullable */
  manual_estimated_minutes: number | null;
  priority: string;
  progress: number;
  /** @nullable */
  project_id: string | null;
  /** @nullable */
  project_name: string | null;
  /** @nullable */
  remaining_days: number | null;
  revision: number;
  status: string;
  title: string;
  updated_at: string;
  user_id: string;
}
// ...
export interface RestActionItemResponse {
  completed?: boolean;
  created_at?: number;
  description?: string;
  due_date?: string;
  /** @nullable */
  estimated_minutes: number | null;
  frequencies?: string[];
  frequency_anchor_date?: string;
  id?: string;
  interval_weeks?: number;
  is_exception?: boolean;
  occurrence_date?: string;
  position?: number;
  priority?: string;
  repeat_state?: string;
  series_id?: string;
  task_id?: string;
  timezone?: string;
  title?: string;
  updated_at?: number;
}